avoiding update or delete projects for a user without permissions

+++ src/controllers/ProjectController.ts
@@ -56,6 +56,10 @@ export class ProjectController {
         const error = new Error('Proyecto no encontrado')
         return res.status(404).json({error: error.message})
       }
+      if(project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error('Sólo el mánager puede actualizar un proyecto')
+        return res.status(404).json({error: error.message})
+      }
       project.projectName = req.body.projectName
       project.clientName = req.body.clientName
       project.description = req.body.description
@@ -70,12 +74,14 @@ export class ProjectController {
     const { id } = req.params
     try {
       const project = await Project.findById(id)
-
       if (!project) {
         const error = new Error('Proyecto no encontrado')
         return res.status(404).json({error: error.message})
       }
-      
+      if(project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error('Sólo el mánager puede eliminar un proyecto')
+        return res.status(404).json({error: error.message})
+      }
       await project.deleteOne()
       res.send('Proyecto eliminado')
     } catch (error) {
